requirements.txt
Flask==2.1.0
gunicorn==20.1.0
psycopg2==2.9.7
python-dotenv==1.0.0